+++ src/video/scalers/Multiply32.cc
@@ -1,6 +1,7 @@
 #include "Multiply32.hh"
 #include "PixelOperations.hh"
 #include "enumerate.hh"
+#include "narrow.hh"
 #include "ranges.hh"
 
 namespace openmsx {
@@ -21,9 +22,9 @@ Multiply32<uint16_t>::Multiply32(const PixelOperations<uint16_t>& pixelOps)
 	Gmask1 = pixelOps.getGmask();
 	Bmask1 = pixelOps.getBmask();
 
-	Rshift1 = ((2 + pixelOps.getRloss()) - pixelOps.getRshift()) & 31;
-	Gshift1 = ((2 + pixelOps.getGloss()) - pixelOps.getGshift()) & 31;
-	Bshift1 = ((2 + pixelOps.getBloss()) - pixelOps.getBshift()) & 31;
+	Rshift1 = narrow<int>(((2 + pixelOps.getRloss()) - pixelOps.getRshift()) & 31);
+	Gshift1 = narrow<int>(((2 + pixelOps.getGloss()) - pixelOps.getGshift()) & 31);
+	Bshift1 = narrow<int>(((2 + pixelOps.getBloss()) - pixelOps.getBshift()) & 31);
 
 	Rmask2 = ((1 << (2 + pixelOps.getRloss())) - 1) <<
 	                (10 + pixelOps.getRshift() - 2 * (2 + pixelOps.getRloss()));
@@ -32,9 +33,9 @@ Multiply32<uint16_t>::Multiply32(const PixelOperations<uint16_t>& pixelOps)
 	Bmask2 = ((1 << (2 + pixelOps.getBloss())) - 1) <<
 	                (10 + pixelOps.getBshift() - 2 * (2 + pixelOps.getBloss()));
 
-	Rshift2 = (2 * (2 + pixelOps.getRloss()) - pixelOps.getRshift() - 10) & 31;
-	Gshift2 = (2 * (2 + pixelOps.getGloss()) - pixelOps.getGshift() - 10) & 31;
-	Bshift2 = (2 * (2 + pixelOps.getBloss()) - pixelOps.getBshift() - 10) & 31;
+	Rshift2 = narrow<int>((2 * (2 + pixelOps.getRloss()) - pixelOps.getRshift() - 10) & 31);
+	Gshift2 = narrow<int>((2 * (2 + pixelOps.getGloss()) - pixelOps.getGshift() - 10) & 31);
+	Bshift2 = narrow<int>((2 * (2 + pixelOps.getBloss()) - pixelOps.getBshift() - 10) & 31);
 
 	Rshift3 = (Rshift1 +  0) & 31;
 	Gshift3 = (Gshift1 + 10) & 31;
